refactor tooltip into tooltipcontent
