+++ src/components/TooltipContent/index.js
@@ -1,0 +1,42 @@
+import PropTypes from "prop-types";
+import React from "react";
+import styled from "styled-components";
+import { Caption } from "../atoms/Typography";
+import Paper from "../Paper/Paper";
+
+const StyledPaper = styled(Paper)`
+  position: relative;
+  width: 272px;
+`;
+
+const Wrapper = styled.div`
+  width: 272px;
+`;
+
+const Title = styled.p`
+  font-size: 1em;
+  margin-bottom: 1em;
+`;
+
+const TooltipContent = ({ title, description }) => {
+  return (
+    <Wrapper>
+      <Title>{title}</Title>
+      <Caption text={description} />
+    </Wrapper>
+  );
+};
+
+TooltipContent.propTypes = {
+  title: PropTypes.string,
+  description: PropTypes.string
+};
+
+TooltipContent.defaultProps = {
+  title: "Title of tooltip",
+  description: "A short description should go here."
+};
+
+TooltipContent.displayName = "Tooltip";
+
+export default TooltipContent;
